rename file to GooglePayProvider
